Test mutually exclusive flags on shrewd querysets.
Assert that a shrewd queryset cannot exist in its shrewd
mode, and be operating on a recycle bin at the same time.

MOTIVATION:
In its shrewd mode a queryset only operates on
active objects which have not yet been 'deleted'
while when operating on a recycle bin it only operates on
objects that have been 'deleted'.