Change default mode of shrewd queryset from shrewd mode.
Previously, the default mode of shrewd querysets was the shrewd mode, in which
only active, non-soft-deleted objects of the underlying model could be fetched.
From now onwards, the default mode of shrewd querysets is the regular
permissive mode of querysets in Django in which every object of the underlying
model can be fetched.

One would have to instantiate a queryset with the flag `shrewd_mode` set to True
in order to have it operate in the shrewd mode.